requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
idna==2.5
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
requests==2.18.3
requests-oauthlib==0.8.0
twython==3.5.0
urllib3==1.22
